fixed the endpoints by adding the corr figs and relative errors and fixed signatures

+++ RootFinding/OpenMethods/secant_method.py
@@ -18,7 +18,7 @@ class SecantSolver :
 
     def solve(self,
               first : float, second : float,  
-              max_itrs : int, tol : float, sig_figs : int) -> tuple[float, int, int, float, float] :
+              max_itrs : int, tol : float, sig_figs : int) -> tuple[float, int, int, float, int] :
         
         errors = []
         for i in range(max_itrs) :
